Resume-Tailor/backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2docx
openai
python-docx
docx2pdf
python-dotenv
